Introducing new cli to get state of daemons
This new cli when functioning entirely should be able to get a
statedump for daemons like glusterd, glusterfsd, etc. This will
be implemented initially to support only glusterd; but can be
extended in the future to accommodate other daemons.

This is not the same as the "gluster volume statedump" cli. This
new cli would provide information about the resources managed by
the daemon. For example for glusterd, the state would consist of
the list of peers, list of volumes, and so on.

Change-Id: I709be84bb9060495f50e328903c2b995de4e2788
Signed-off-by: Samikshan Bairagya <samikshan@gmail.com>