cli, glusterd: New CLI command to get maximum supported op-version
gluster volume get <VOLNAME> cluster.opversion gives us the current
op-version on which the cluster is operating. There is no command
that lets the user know the maximum supported op-version that the
cluster can run on. This new CLI command introduces that.

Usage:
 # gluster get-max-opversion

Output:
Maximum supported op-version: 30800

NOTE: The only way to test this feature for now is to set the
GD_OP_VERSION_MAX macro to different values (30800 for 3.8, 30900
for 3.9, and so on) and rebuild glusterd. Since the regression test
framework currently doesn't have support to simulate these tests,
there are no accompanying regression tests for this feature. It
should be possible to add tests once glusto comes in and makes
it easier to run a heterogeneous cluster through the regression
testing framework.

Change-Id: I379803528a641cd7dd616c3a6de283f1aa3f7c95
Signed-off-by: Samikshan Bairagya <samikshan@gmail.com>