glusterd: Log option retrieval failures from xlators as warnings
While trying to get option value through volume get functionality,
if the option value is not set in the respective translator volume
options table, it is logged in glusterd log as an error, instead of
a warning. This patch changes that behaviour and logs these option
retrieval failures as warnings.

Change-Id: Ied0b9c8f5daee6b25b4eeecd652862bd44f78165
Signed-off-by: Samikshan Bairagya <samikshan@gmail.com>