core: run many bricks within one glusterfsd process
This patch adds support for multiple brick translator stacks running
in a single process.  It does this by defining a new RPC
(GLUSTERD_BRICK_ATTACH) which can be sent to a glusterfsd process
specifying a volfile path.  Upon receipt, that process will build a
graph (minus the protocol/server translator at the top) and attach it
under its existing server translator.  There is a corresponding detach
RPC for when a brick needs to go away.  Both are issued automatically
from glusterd at the appropriate times - volume start/stop, brick
add/remove, etc.  There is also a standalone utility to issue these
RPCs for testing or other purposes.

There are also changes to move various things from the server
translator or its first child to the "correct" brick stack for
whatever's going on.  For example, authentication options are
automatically moved down from where they're defined on the server
translator (by volgen) to the top of each brick stack, so that each
brick can authenticate differently.  Notify and I/O path code pass
control to the brick associated with the connection that triggered
them, and so on.

This version reliably passes all of the tests in tests/basic except
that those that are already marked bad (actually it passes some of
them too), with one caveat.  Changelog and trash are disabled, because
in different ways each is incompatible with multiplexing.  This will
be fixed in a future version of the patch.

In some cases, getting these tests to pass required changes to the
tests because those made bad assumptions - especially about bricks
becoming available immediately after a volume is started.  Also, some
tests will take so long to run that they might as well be considered
to have failed, unless the new mem-pool code is also present.

Even though the code is present, multiplexing is not currently done
(i.e. every brick still gets its own process) unless the MULTIPLEX
environment variable is defined.  This makes it easy to test both
multiplexing and non-multiplexing modes of operation, compare
performance between them, etc., without having to recompile.

Change-Id: I45059454e51d6f4cbb29a4953359c09a408695cb
BUG: 1385758
Signed-off-by: Jeff Darcy <jdarcy@redhat.com>